R implementation of the LEfSe method

Perform a LEfSe analysis: the function carries out differential analysis between two sample groups for multiple microorganisms and uses linear discirminant analysis to establish their effect sizes. Subclass information for each class can be incorporated into the analysis (see examples). Microorganisms with large differences between two sample groups are identified as biomarkers.

lefser(
  expr,
  kruskal.threshold = 0.05,
  wilcox.threshold = 0.05,
  lda.threshold = 2,
  groupCol = "GROUP",
  blockCol = NULL,
  assay = 1L,
  trim.names = FALSE
)

Arguments

expr
A SummarizedExperiment with expression data.

kruskal.threshold
numeric(1) The p-value for the Kruskal-Wallis Rank Sum Test (default 0.05).

wilcox.threshold
numeric(1) The p-value for the Wilcoxon Rank-Sum Test when 'blockCol' is present (default 0.05).

lda.threshold
numeric(1) The effect size threshold (default 2.0).

groupCol
character(1) Column name in `colData(expr)` indicating groups, usually a factor with two levels (e.g., `c("cases", "controls")`; default "GROUP").

blockCol
character(1) Optional column name in `colData(expr)` indicating the blocks, usually a factor with two levels (e.g., `c("adult", "senior")`; default NULL).

assay
The i-th assay matrix in the `SummarizedExperiment` ('expr'; default 1).

trim.names
If `TRUE` extracts the most specific taxonomic rank of organism.

Value

The function returns a dataframe with two columns, which are names of microorganisms and their LDA scores.

Examples

    # (1) Using classes only
    data(zeller14)
    # exclude 'adenoma'
    zeller14 <- zeller14[, zeller14$study_condition != "adenoma"]
    res_group <- lefser(zeller14, groupCol = "study_condition")
    head(res_group)

#>                                                                  Names
#> 1                                            k__Bacteria|p__Firmicutes
#> 2                              k__Bacteria|p__Firmicutes|c__Clostridia
#> 3             k__Bacteria|p__Firmicutes|c__Clostridia|o__Clostridiales
#> 4                                        k__Bacteria|p__Actinobacteria
#> 5                      k__Bacteria|p__Actinobacteria|c__Actinobacteria
#> 6 k__Bacteria|p__Actinobacteria|c__Actinobacteria|o__Bifidobacteriales
#>      scores
#> 1 -6.536708
#> 2 -6.458054
#> 3 -6.458054
#> 4 -6.282184
#> 5 -6.282184
#> 6 -6.237298

    # (2) Using classes and sublasses
    data(zeller14)
    # exclude 'adenoma'
    zeller14 <- zeller14[, zeller14$study_condition != "adenoma"]
    res_block <- lefser(
         zeller14, groupCol = "study_condition", blockCol = "age_category"
    )
    head(res_block)

#>                                                                                                                                        Names
#> 1                                                k__Bacteria|p__Firmicutes|c__Clostridia|o__Clostridiales|f__Ruminococcaceae|g__Ruminococcus
#> 2                                                                                    k__Bacteria|p__Firmicutes|c__Bacilli|o__Lactobacillales
#> 3                                                                                                       k__Bacteria|p__Firmicutes|c__Bacilli
#> 4                  k__Bacteria|p__Firmicutes|c__Clostridia|o__Clostridiales|f__Ruminococcaceae|g__Ruminococcus|s__Ruminococcus_sp_5_1_39BFAA
#> 5 k__Bacteria|p__Firmicutes|c__Clostridia|o__Clostridiales|f__Ruminococcaceae|g__Ruminococcus|s__Ruminococcus_sp_5_1_39BFAA|t__GCF_000159975
#> 6                            k__Bacteria|p__Firmicutes|c__Clostridia|o__Clostridiales|f__Eubacteriaceae|g__Eubacterium|s__Eubacterium_hallii
#>      scores
#> 1 -6.172329
#> 2 -5.914480
#> 3 -5.904209
#> 4 -5.520557
#> 5 -5.520557
#> 6 -5.475524