R implementation of the LEfSe method

Perform a LEfSe analysis: the function carries out differential analysis between two sample groups for multiple microorganisms and uses linear discirminant analysis to establish their effect sizes. Subclass information for each class can be incorporated into the analysis (see examples). Microorganisms with large differences between two sample groups are identified as biomarkers.

lefser(
  expr,
  kw.threshold = 0.05,
  wilcoxon.threshold = 0.05,
  lda.threshold = 2,
  groupCol = "GROUP",
  blockCol = "BLOCK",
  assay = 1L
)

Arguments

expr
A SummarizedExperiment with expression data.

kw.threshold
The p-value threshold for Kruskal-Wallis Rank Sum Test. The default is at <= 0.05.

wilcoxon.threshold
The p-value for Wilcoxon Rank-Sum Test. The default is at <= 0.05.

lda.threshold
The effect size threshold. The default is at 2.0.

groupCol
character(1) Column name in `colData(expr)` indicating groups, usually a factor with two levels (e.g., `c("cases", "controls")`).

blockCol
character(1) Column name in `colData(expr)` indicating the blocks, usually a factor with two levels (e.g., `c("adult", "senior")`).

assay
The i-th assay matrix in the `SummarizedExperiment` ('expr'). Defaults to 1.

Value

The function returns a dataframe with two columns, which are names of microorganisms and their LDA scores.

Examples

    # (1) Using classes only
    data(zeller14)
    # exclude 'adenoma'
    zeller14 <- zeller14[, zeller14$study_condition != "adenoma"]
    # assign '0' class to 'conrol' and '1' to 'CRC' (i.e., colorectal cancer)
    zeller14$GROUP <- ifelse(zeller14$study_condition == "control", 0, 1)
    results <- lefser(zeller14)

#> Length of block: 0

    head(results)

#>                  Names    scores
#> 1        p__Firmicutes -6.536708
#> 2        c__Clostridia -6.458054
#> 3     o__Clostridiales -6.458054
#> 4    p__Actinobacteria -6.282184
#> 5    c__Actinobacteria -6.282184
#> 6 o__Bifidobacteriales -6.237298

    # (2) Using classes and sublasses
    data(zeller14)
    # exclude 'adenoma'
    zeller14 <- zeller14[, zeller14$study_condition != "adenoma"]
    # assign '0' class to 'conrol' and '1' to 'CRC' (i.e., colorectal cancer)
    zeller14$GROUP <- ifelse(zeller14$study_condition == "control", 0, 1)
    # assign '0' class to 'adult' and '1' to 'senior'
    zeller14$BLOCK <- ifelse(zeller14$age_category == "adult", 0, 1)
    results <- lefser(zeller14)

#> Length of block: 157

#> Warning: The conditional covariance matrix has zero diagonal elements

#> Warning: The conditional covariance matrix has zero diagonal elements

#> Warning: The conditional covariance matrix has zero diagonal elements

#> Warning: The conditional covariance matrix has zero diagonal elements

#> Warning: The conditional covariance matrix has zero diagonal elements

#> Warning: The conditional covariance matrix has zero diagonal elements

#> Warning: The conditional covariance matrix has zero diagonal elements

#> Warning: The conditional covariance matrix has zero diagonal elements

#> Warning: The conditional covariance matrix has zero diagonal elements

#> Warning: The conditional covariance matrix has zero diagonal elements

    head(results)

#>                           Names    scores
#> 1            o__Lactobacillales -5.937486
#> 2 s__Ruminococcus_sp_5_1_39BFAA -5.630104
#> 3              t__GCF_000159975 -5.630104
#> 4         s__Eubacterium_hallii -5.493462
#> 5              t__GCF_000173975 -5.493462
#> 6   s__Streptococcus_salivarius -5.353776